fix(adapters): pomegranate int evidence + gpytorch as_tensor (§A.3, §A.8)
Bug 3 (pomegranate): predict_proba indexes per-node probability tensors
with the evidence values; passing float trips
``IndexError: tensors used as indices must be long, …``.  The adapter
is discrete-only, so the masked-tensor `row` is now `dtype=long` and
each evidence value is cast to `int` before insertion.

Bug 8 (gpytorch): swap ``torch.tensor(some_tensor)`` for
``torch.as_tensor(...)`` so we don't trip the ``UserWarning`` to
``detach().clone()`` when ``q.evidence`` values arrive as tensors
already.

+++ benchmarking/baselines/pomegranate_adapter.py
@@ -109,17 +109,20 @@ class PomegranateAdapter(BaselineAdapter):
 
         target = q.targets[0]
         n = len(self._topo)
-        row = torch.zeros((1, n))
+        # PR-B §A.3: pomegranate v1.1's predict_proba uses evidence values
+        # as indices into per-node probability tensors; if `row` is float
+        # we hit ``IndexError: tensors used as indices must be long…``.
+        # The adapter is discrete-only, so all evidence values can safely
+        # be encoded as long.
+        row = torch.zeros((1, n), dtype=torch.long)
         mask = torch.zeros((1, n), dtype=torch.bool)
         for k, v in q.evidence.items():
             i = self._node_to_idx[k]
-            # Reduce any tensor (incl. shape [1] or [B] when called per-row)
-            # to a scalar Python int.
             if isinstance(v, torch.Tensor):
                 scalar = int(v.reshape(-1)[0].item())
             else:
                 scalar = int(v)
-            row[0, i] = float(scalar)
+            row[0, i] = scalar
             mask[0, i] = True
         with warnings.catch_warnings():
             warnings.simplefilter("ignore", UserWarning)
